sys/shell/cmds: add cat and tee cmds
This patch moves these shell cmds out of the filesystem example into the
shell cmds folder. This is done to make them available to all
applications.
