+++ examples/filesystem/main.c
@@ -19,86 +19,9 @@
  */
 
 #include <stdio.h>
-#include <string.h>
-#include <unistd.h>
-#include <fcntl.h>
 
-#include "shell.h"
-
-static int _cat(int argc, char **argv)
-{
-    if (argc < 2) {
-        printf("Usage: %s <file>\n", argv[0]);
-        return 1;
-    }
-    /* With newlib or picolibc, low-level syscalls are plugged to RIOT vfs
-     * on native, open/read/write/close/... are plugged to RIOT vfs */
-#if defined(MODULE_NEWLIB) || defined(MODULE_PICOLIBC)
-    FILE *f = fopen(argv[1], "r");
-    if (f == NULL) {
-        printf("file %s does not exist\n", argv[1]);
-        return 1;
-    }
-    char c;
-    while (fread(&c, 1, 1, f) != 0) {
-        putchar(c);
-    }
-    fclose(f);
-#else
-    int fd = open(argv[1], O_RDONLY);
-    if (fd < 0) {
-        printf("file %s does not exist\n", argv[1]);
-        return 1;
-    }
-    char c;
-    while (read(fd, &c, 1) != 0) {
-        putchar(c);
-    }
-    close(fd);
-#endif
-    fflush(stdout);
-    return 0;
-}
-
-static int _tee(int argc, char **argv)
-{
-    if (argc != 3) {
-        printf("Usage: %s <file> <str>\n", argv[0]);
-        return 1;
-    }
-
-#if defined(MODULE_NEWLIB) || defined(MODULE_PICOLIBC)
-    FILE *f = fopen(argv[1], "w+");
-    if (f == NULL) {
-        printf("error while trying to create %s\n", argv[1]);
-        return 1;
-    }
-    if (fwrite(argv[2], 1, strlen(argv[2]), f) != strlen(argv[2])) {
-        puts("Error while writing");
-    }
-    fclose(f);
-#else
-    int fd = open(argv[1], O_RDWR | O_CREAT, 00777);
-    if (fd < 0) {
-        printf("error while trying to create %s\n", argv[1]);
-        return 1;
-    }
-    if (write(fd, argv[2], strlen(argv[2])) != (ssize_t)strlen(argv[2])) {
-        puts("Error while writing");
-    }
-    close(fd);
-#endif
-    return 0;
-}
-
-static const shell_command_t shell_commands[] = {
-    { "cat", "print the content of a file", _cat },
-    { "tee", "write a string in a file", _tee },
-    { NULL, NULL, NULL }
-};
-
-/* constfs example */
 #include "fs/constfs.h"
+#include "shell.h"
 
 #define HELLO_WORLD_CONTENT "Hello World!\n"
 #define HELLO_RIOT_CONTENT  "Hello RIOT!\n"
@@ -142,7 +65,7 @@ int main(void)
     }
 
     char line_buf[SHELL_DEFAULT_BUFSIZE];
-    shell_run(shell_commands, line_buf, SHELL_DEFAULT_BUFSIZE);
+    shell_run(NULL, line_buf, SHELL_DEFAULT_BUFSIZE);
 
     return 0;
 }
